[DEV-6496] Moves DunsCount into list_recipients file to reduce code duplication
